requirements.txt
beautifulsoup4
pyppeteer
flask[async]
python-dotenv
boto3
datetime
requests
Pillow
sentencepiece
transformers
replicate